dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Local developer dashboard for managing project repos and Docker containers"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "flask>=3",
    "flask-cors>=4",
    "pyyaml>=6",
]

[project.optional-dependencies]
tui = [
    "textual>=0.80",
    "rich>=13",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
dashboard = "dashboard.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dashboard = ["templates/**/*", "static/**/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 300

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S108", "S603", "S607", "S310"]
"src/dashboard/web.py" = ["T201", "S603", "S607", "S104"]
"src/dashboard/__main__.py" = ["T201"]
"src/dashboard/scanner.py" = ["S603", "S607", "S110", "SIM102", "SIM105"]
"src/dashboard/process_manager.py" = ["S104", "S110", "RUF006"]
"src/dashboard/app.py" = ["S110", "RUF012"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "textual.*"
ignore_missing_imports = true

[tool.coverage.run]
source = ["dashboard"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__",
]
